parity commit: lmstudio.js #86 magic numbers
