+++ planning/todos.md
@@ -5,10 +5,7 @@ TODO: unit tests!
 TODO: single vs double underscore naming
 TODO: subclasses of clientport with restrictions on request schemas?
 TODO: clean up streamable iterators
-
 TODO: sync code does not respect SIGINT
-TODO: code hangs on RPC errors
-TODO: make return types Optional for error handling
-TODO: clean up import statements, alphabetize
-TODO: type hints, abstractmethods
+TODO: type hints
 TODO: README + docs
+TODO: preprocessors
