+++ src/lmstudio_sdk/utils/__init__.py
@@ -3,7 +3,7 @@ from .BufferedEvent import BufferedEvent
 from .logger import get_logger, RECV, SEND, WRAPPER, WEBSOCKET
 from .PseudoFuture import PseudoFuture
 from .sync_async_decorator import pretty_print, pretty_print_error, sync_async_decorator
-from .utils import ChannelError, lms_default_ports, generate_random_base64, RPCError
+from .utils import ChannelError, lms_default_ports, generate_random_base64, number_to_checkbox_numeric, RPCError
 
 __all__ = [
     "AbortSignal",
@@ -12,6 +12,7 @@ __all__ = [
     "generate_random_base64",
     "get_logger",
     "lms_default_ports",
+    "number_to_checkbox_numeric",
     "pretty_print",
     "pretty_print_error",
     "PseudoFuture",
